requirements.txt
filelock==3.19.1
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
PyYAML==6.0.3
shiboken6==6.10.1
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
typing_extensions==4.15.0
